refactor: :truck: Renomeação da pasta `@types` para `schemas` e alteração do `UserProps`
- Renomeação da pasta `@types` para `schemas` e de `index.ts` para `UserProps.ts`
- Alteração do `UserProps` para seguir o padrão de `schemas` com `Zod`
- Ajuste das importações de `UserProps`
