--- src/shared/@types/index.ts
@@ -1,6 +1,0 @@
-export type UserProps = {
-	avatar: string
-	email: string
-	id: string
-	name: string
-}
